Package options added + part problem header fix
Added all package options as requested in #23, so this fixes #23.

Further now part problems show exercise numbers if they are
the first part problem on a new page. If a problem is declared
on the current page, the part problem will not show problem number
(see docs for more details). Added everypage.tex for this purpose.